Move Galaxy to AreaItems folder
