+++ PyRoute/AreaItems/Galaxy.py
