+++ Tests/testTradeCalculation.py
@@ -5,9 +5,9 @@ import sys
 from PyRoute.Calculation.TradeCalculation import TradeCalculation
 
 sys.path.append('../PyRoute')
+from PyRoute.AreaItems.Galaxy import Galaxy
 from PyRoute.AreaItems.Sector import Sector
 from PyRoute.Star import Star
-from PyRoute.Galaxy import Galaxy
 
 
 class testTradeCalculation(unittest.TestCase):
